+++ Facebook/src/main/java/com/applovin/mediation/adapters/FacebookMediationAdapter.java
@@ -578,6 +578,18 @@ public class FacebookMediationAdapter
         return "APPLOVIN_" + AppLovinSdk.VERSION + ":" + getAdapterVersion();
     }
 
+    private MaxNativeAdView createMaxNativeAdView(final MaxNativeAd maxNativeAd, final String templateName, final Activity activity)
+    {
+        if ( AppLovinSdk.VERSION_CODE >= 11010000 )
+        {
+            return new MaxNativeAdView( maxNativeAd, templateName, getApplicationContext() );
+        }
+        else
+        {
+            return new MaxNativeAdView( maxNativeAd, templateName, activity );
+        }
+    }
+
     private class InterstitialAdListener
             implements InterstitialAdExtendedListener
     {
@@ -903,7 +915,8 @@ public class FacebookMediationAdapter
 
                     final MaxNativeAd maxNativeAd = new MaxNativeAd.Builder()
                             .setAdFormat( adFormat )
-                            .setTitle( mNativeAdViewAd.getAdvertiserName() )
+                            .setTitle( mNativeAdViewAd.getAdHeadline() )
+                            .setAdvertiser( mNativeAdViewAd.getAdvertiserName() )
                             .setBody( mNativeAdViewAd.getAdBodyText() )
                             .setCallToAction( mNativeAdViewAd.getAdCallToAction() )
                             .setIconView( iconView )
@@ -924,20 +937,24 @@ public class FacebookMediationAdapter
                         if ( templateName.equals( "vertical" ) )
                         {
                             String verticalTemplateName = ( adFormat == MaxAdFormat.LEADER ) ? "vertical_leader_template" : "vertical_media_banner_template";
-                            maxNativeAdView = new MaxNativeAdView( maxNativeAd, verticalTemplateName, activity );
+                            maxNativeAdView = createMaxNativeAdView( maxNativeAd, verticalTemplateName, activity );
                         }
                         else
                         {
-                            maxNativeAdView = new MaxNativeAdView( maxNativeAd, templateName, activity );
+                            maxNativeAdView = createMaxNativeAdView( maxNativeAd, templateName, activity );
                         }
                     }
                     else if ( AppLovinSdk.VERSION_CODE < 9140500 )
                     {
-                        maxNativeAdView = new MaxNativeAdView( maxNativeAd, AppLovinSdkUtils.isValidString( templateName ) ? templateName : "no_body_banner_template", activity );
+                        maxNativeAdView = createMaxNativeAdView( maxNativeAd,
+                                                                 AppLovinSdkUtils.isValidString( templateName ) ? templateName : "no_body_banner_template",
+                                                                 activity );
                     }
                     else
                     {
-                        maxNativeAdView = new MaxNativeAdView( maxNativeAd, AppLovinSdkUtils.isValidString( templateName ) ? templateName : "media_banner_template", activity );
+                        maxNativeAdView = createMaxNativeAdView( maxNativeAd,
+                                                                 AppLovinSdkUtils.isValidString( templateName ) ? templateName : "media_banner_template",
+                                                                 activity );
                     }
 
                     final List<View> clickableViews = new ArrayList<>();
@@ -958,6 +975,10 @@ public class FacebookMediationAdapter
                     if ( AppLovinSdkUtils.isValidString( maxNativeAd.getCallToAction() ) && maxNativeAdView.getCallToActionButton() != null )
                     {
                         clickableViews.add( maxNativeAdView.getCallToActionButton() );
+                    }
+                    if ( AppLovinSdkUtils.isValidString( maxNativeAd.getAdvertiser() ) && maxNativeAdView.getAdvertiserTextView() != null )
+                    {
+                        clickableViews.add( maxNativeAdView.getAdvertiserTextView() );
                     }
                     if ( AppLovinSdkUtils.isValidString( maxNativeAd.getBody() ) && maxNativeAdView.getBodyTextView() != null )
                     {
@@ -1130,7 +1151,8 @@ public class FacebookMediationAdapter
         {
             final MaxFacebookNativeAd maxNativeAd = new MaxFacebookNativeAd( new MaxNativeAd.Builder()
                                                                                      .setAdFormat( MaxAdFormat.NATIVE )
-                                                                                     .setTitle( nativeAd.getAdvertiserName() )
+                                                                                     .setTitle( nativeAd.getAdHeadline() )
+                                                                                     .setAdvertiser( nativeAd.getAdvertiserName() )
                                                                                      .setBody( nativeAd.getAdBodyText() )
                                                                                      .setCallToAction( nativeAd.getAdCallToAction() )
                                                                                      .setIcon( new MaxNativeAd.MaxNativeAdImage( iconDrawable ) )
@@ -1143,12 +1165,12 @@ public class FacebookMediationAdapter
         {
             if ( isTemplateAd )
             {
-                return AppLovinSdkUtils.isValidString( nativeAd.getAdvertiserName() );
+                return AppLovinSdkUtils.isValidString( nativeAd.getAdHeadline() );
             }
             else
             {
                 // NOTE: media view is created and will always be non-null
-                return AppLovinSdkUtils.isValidString( nativeAd.getAdvertiserName() )
+                return AppLovinSdkUtils.isValidString( nativeAd.getAdHeadline() )
                         && AppLovinSdkUtils.isValidString( nativeAd.getAdCallToAction() );
             }
         }
@@ -1176,6 +1198,10 @@ public class FacebookMediationAdapter
             if ( AppLovinSdkUtils.isValidString( getTitle() ) && maxNativeAdView.getTitleTextView() != null )
             {
                 clickableViews.add( maxNativeAdView.getTitleTextView() );
+            }
+            if ( AppLovinSdkUtils.isValidString( getAdvertiser() ) && maxNativeAdView.getAdvertiserTextView() != null )
+            {
+                clickableViews.add( maxNativeAdView.getAdvertiserTextView() );
             }
             if ( AppLovinSdkUtils.isValidString( getBody() ) && maxNativeAdView.getBodyTextView() != null )
             {