+++ Facebook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.8.0.7
+* Add support for new `MaxNativeAdView` constructor taking in non-Activity context.
+  
+## 6.8.0.6
+* Fix headline and advertiser views in native ads.
+
 ## 6.8.0.5
 * Fix native ad icon not rendering.